#33 Updating API endpoints
Removing duplicate collection attributes
