+++ src/app/models.py
@@ -1,3 +1,4 @@
+from itertools import chain
 from typing import TypeVar
 
 from app.elastic.fields import Field, FieldType
@@ -20,3 +21,20 @@ class ElasticResourceAttribute(Field):
     EDU_CONTEXT = ("properties.ccm:educationalcontext", FieldType.TEXT)
     EDU_CONTEXT_DE = ("i18n.de_DE.ccm:educationalcontext", FieldType.TEXT)
     REPLICATION_SOURCE_DE = ("replicationsource", FieldType.TEXT)
+
+
+class _CollectionAttribute(Field):
+    TITLE = ("properties.cm:title", FieldType.TEXT)
+    DESCRIPTION = ("properties.cm:description", FieldType.TEXT)
+    PATH = ("path", FieldType.KEYWORD)
+    PARENT_ID = ("parentRef.id", FieldType.KEYWORD)
+    NODE_ID = ("nodeRef.id", FieldType.KEYWORD)
+
+
+CollectionAttribute = Field(
+    "CollectionAttribute",
+    [
+        (f.name, (f.value, f.field_type))
+        for f in chain(ElasticResourceAttribute, _CollectionAttribute)
+    ],
+)
